Allow a map to be set to readonly
In certain use cases we want the use to be able to interact with the map without moving the pin.

For example, displaying an address input & showing a map for the corresponding lookup address. Or manually editing the lat & lng fields.

I am using the library to allow entry of an address and have that address be displayed on a map (with lat & lng values stored for later use). In this instance it is not ideal to have to pin move if the user clicks on the map (which they may do to zoom/pan etc).